# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AccordBridge workspace API",
    "version": "0.1.0",
    "description": "Development workspace. No payment endpoints. Sessions use HttpOnly cookies; unsafe requests require an exact allowed Origin and X-AccordBridge-Request: 1."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4000/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accordbridge_session"
      }
    },
    "schemas": {
      "Register": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "Login": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "summary": "Create an account and session",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Must equal configured FRONTEND_ORIGIN"
          },
          {
            "in": "header",
            "name": "X-AccordBridge-Request",
            "required": true,
            "schema": {
              "const": "1"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Origin/header or action forbidden"
          },
          "404": {
            "description": "Record not found or not accessible"
          },
          "409": {
            "description": "Stale version/draft, locked terms or duplicate account"
          },
          "429": {
            "description": "Too many requests"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Sign in and rotate the browser session",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Must equal configured FRONTEND_ORIGIN"
          },
          {
            "in": "header",
            "name": "X-AccordBridge-Request",
            "required": true,
            "schema": {
              "const": "1"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Origin/header or action forbidden"
          },
          "404": {
            "description": "Record not found or not accessible"
          },
          "409": {
            "description": "Stale version/draft, locked terms or duplicate account"
          },
          "429": {
            "description": "Too many requests"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Revoke session and clear cookie",
        "security": [],
        "parameters": [
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Must equal configured FRONTEND_ORIGIN"
          },
          {
            "in": "header",
            "name": "X-AccordBridge-Request",
            "required": true,
            "schema": {
              "const": "1"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Origin/header or action forbidden"
          },
          "404": {
            "description": "Record not found or not accessible"
          },
          "409": {
            "description": "Stale version/draft, locked terms or duplicate account"
          },
          "429": {
            "description": "Too many requests"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Read the signed-in account",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid fields"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Origin/header or action forbidden"
          },
          "404": {
            "description": "Record not found or not accessible"
          },
          "409": {
            "description": "Stale version/draft, locked terms or duplicate account"
          },
          "429": {
            "description": "Too many requests"
          }
        }
      }
    },
...
  }
}